+++ src/TransformComponent.hpp
@@ -1,11 +1,12 @@
 #pragma once
 
+#include "Matrix4x4.hpp"
 #include "Vector3.hpp"
 #include "Quaternion.hpp"
 
 struct TransformComponent {
     int entity_id;
-    Vector3d position;
-    Quaterniond rotation;
-    Vector3d scale;
+    Vector3f position;
+    Quaternionf rotation;
+    Vector3f scale;
 };
